# Speedup insert performance of vector similarity index by parallelization