--- a/src/Interpreters/Context.cpp
+++ b/src/Interpreters/Context.cpp
@@ -9,6 +9,7 @@
 #include <Common/SensitiveDataMasker.h>
 #include <Common/Macros.h>
 #include <Common/EventNotifier.h>
+#include <Common/getNumberOfPhysicalCPUCores.h>
 #include <Common/Stopwatch.h>
 #include <Common/formatReadable.h>
 #include <Common/Throttler.h>
@@ -120,7 +121,6 @@
 #include <Interpreters/InterpreterSelectWithUnionQuery.h>
 #include <base/defines.h>
 
-
 namespace fs = std::filesystem;
 
 namespace ProfileEvents
@@ -168,6 +168,9 @@ namespace CurrentMetrics
     extern const Metric AttachedDictionary;
     extern const Metric AttachedDatabase;
     extern const Metric PartsActive;
+    extern const Metric UsearchUpdateThreads;
+    extern const Metric UsearchUpdateThreadsActive;
+    extern const Metric UsearchUpdateThreadsScheduled;
 }
 
 
@@ -296,6 +299,8 @@ struct ContextSharedPart : boost::noncopyable
     mutable std::unique_ptr<ThreadPool> load_marks_threadpool;  /// Threadpool for loading marks cache.
     mutable OnceFlag prefetch_threadpool_initialized;
     mutable std::unique_ptr<ThreadPool> prefetch_threadpool;    /// Threadpool for loading marks cache.
+    mutable OnceFlag vector_similarity_index_creation_threadpool_initialized;
+    mutable std::unique_ptr<ThreadPool> vector_similarity_index_creation_threadpool; /// Threadpool for vector-similarity index creation.
     mutable UncompressedCachePtr index_uncompressed_cache TSA_GUARDED_BY(mutex);      /// The cache of decompressed blocks for MergeTree indices.
     mutable QueryCachePtr query_cache TSA_GUARDED_BY(mutex);                          /// Cache of query results.
     mutable MarkCachePtr index_mark_cache TSA_GUARDED_BY(mutex);                      /// Cache of marks in compressed files of MergeTree indices.
@@ -3093,6 +3098,25 @@ ThreadPool & Context::getLoadMarksThreadpool() const
     });
 
     return *shared->load_marks_threadpool;
+}
+
+ThreadPool & Context::getVectorSimilarityIndexCreationThreadPool() const
+{
+    callOnce(
+        shared->vector_similarity_index_creation_threadpool_initialized,
+        [&]
+        {
+            const auto & server_setting = getServerSettings();
+            size_t max_thread_pool_size = server_setting.max_thread_pool_size_for_vector_similarity_index_creation > 0
+                ? server_setting.max_thread_pool_size_for_vector_similarity_index_creation
+                : getNumberOfPhysicalCPUCores();
+            shared->vector_similarity_index_creation_threadpool = std::make_unique<ThreadPool>(
+                CurrentMetrics::UsearchUpdateThreads,
+                CurrentMetrics::UsearchUpdateThreadsActive,
+                CurrentMetrics::UsearchUpdateThreadsScheduled,
+                max_thread_pool_size);
+        });
+    return *shared->vector_similarity_index_creation_threadpool;
 }
 
 void Context::setIndexUncompressedCache(const String & cache_policy, size_t max_size_in_bytes, double size_ratio)