### Changelog category (leave one):

- Performance Improvement



### Changelog entry (a user-readable short description of the changes that goes to CHANGELOG.md):
Speedup insert performance with vector similarity index by adding data to vector index parallel


Insert 50k rows vector with 768 dimension:

Before:
```
50000 rows in set. Elapsed: 107.843 sec. Processed 50.00 thousand rows, 482.50 MB (463.64 rows/s., 4.47 MB/s.) Peak memory usage: 338.34 MiB.
```

After:
```
50000 rows in set. Elapsed: 12.377 sec. Processed 50.00 thousand rows, 482.50 MB (4.04 thousand rows/s., 38.98 MB/s.) Peak memory usage: 338.34 MiB.
```