@ucasfl 2024-09-11
>The previous optimization had the problem that it spawned N threads per each merge. With M merges going on at the same time, one had N x M extra threads. Is it right that with your optimization, the number of threads for Usearch index creation is globally fixed?

It could be globally fixed if we create a global thread pool for Usearch index creation.