@rschu1ze 2024-09-11
> It could be globally fixed if we create a global thread pool for Usearch index creation.

Let's do that.