@ucasfl 2024-09-12
> > It could be globally fixed if we create a global thread pool for Usearch index creation.
> 
> Let's do that.

The pool size should be a server setting or something else?